src/requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
alembic
amqp
annotated-types
anyio
appnope
APScheduler
argon2-cffi
argon2-cffi-bindings
arrow
asgiref
asteval
asttokens
async-lru
async-timeout
attrs
babel
backcall
bcrypt
beautifulsoup4
billiard
bleach
blinker
boto3
botocore
cachelib
cachetools
celery
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
comm
contourpy
cron-descriptor
cryptography
cycler
dataclasses-json
debugpy
decorator
defusedxml
distro
Django
django-timezone-field
dnspython
ecdsa
eventlet
executing
fastapi
fastapi-pagination
fastjsonschema
Flask
Flask-Admin
Flask-Caching
Flask-CKEditor
flask-cors
Flask-Migrate
Flask-SQLAlchemy
fonttools
fqdn
frozenlist
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
googletrans
greenlet
h11
h2
hiredis
hpack
httpcore
httplib2
httptools
httpx
hubspot-api-client
hyperframe
idna
iniconfig
ipykernel
ipython
ipython-genutils
ipython_pygments_lexers
isoduration
itsdangerous
jedi
Jinja2
jiter
jmespath
json5
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
kiwisolver
kombu
langchain
langchain-core
langchain-text-splitters
langsmith
loguru
Mako
MarkupSafe
marshmallow
marshmallow-enum
matplotlib
matplotlib-inline
mistune
motor
multidict
mypy-extensions
nbclassic
nbclient
nbconvert
nbformat
nest-asyncio
networkx
notebook
notebook_shim
numpy
oauthlib
ollama
openai
orjson
outcome
overrides
packaging
pandas
pandocfilters
parso
passlib
pexpect
pickleshare
pillow
pinecone-client
pinecone-plugin-inference
pinecone-plugin-interface
platformdirs
pluggy
prometheus_client
prompt_toolkit
propcache
proto-plus
protobuf
psutil
psycopg2-binary
ptyprocess
pure_eval
pyarrow
pyasn1
pyasn1_modules
pycparser
pycurl
pydantic
pydantic_core
Pygments
PyJWT
pymongo
pymongo-auth-aws
pymongoarrow
pymongocrypt
pyparsing
pyrsistent
pytest
python-crontab
python-dateutil
python-dotenv
python-jose
python-json-logger
python-multipart
pytz
PyYAML
pyzmq
qtconsole
QtPy
redis
referencing
requests
requests-oauthlib
requests-toolbelt
rfc3339-validator
rfc3986-validator
rpds-py
rsa
s3transfer
Send2Trash
setuptools
six
sniffio
sortedcontainers
soupsieve
SQLAlchemy
sqlparse
stack-data
starlette
stripe
tenacity
terminado
tinycss2
tornado
tqdm
traitlets
trio
types-python-dateutil
typing-inspect
typing-inspection
typing_extensions
tzdata
tzlocal
uri-template
uritemplate
urllib3
uvicorn
uvloop
vine
watchfiles
wcwidth
webcolors
webencodings
websocket-client
websockets
Werkzeug
widgetsnbextension
WTForms
yarl
zstandard
